+++ lib/widgets/flutter_json_schema_field.dart
@@ -1,9 +1,13 @@
 library flutter_json_schema_form;
 
 import 'dart:convert';
+import 'dart:html';
+import 'dart:js';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:flutter_json_schema_form/controller/flutter_json_schema_form_controller.dart';
 import 'package:flutter_json_schema_form/flutter_json_schema_form.dart';
 import 'package:json_schema_document/json_schema_document.dart';
@@ -20,6 +24,9 @@ class FlutterJsonSchemaFormField extends StatelessWidget {
     this.fileWidget,
     this.formState = const {},
     this.onChanged,
+    this.translateField = false,
+    this.translateBase = '',
+    this.fieldName = ''
   }) : super(key: key);
 
   final JsonSchema jsonSchema;
@@ -34,7 +41,15 @@ class FlutterJsonSchemaFormField extends StatelessWidget {
 
   final Widget? fileWidget;
 
-  String? get title => jsonSchema.title;
+  String? get title => jsonSchema.title ?? '';
+
+  getTitleTranslation(context, key) {
+    if (translateField) {
+      var translateKey = jsonSchema.title != null ? jsonSchema.title.toString() : translateBase.isNotEmpty ? translateBase + '.properties.' + fieldName + '.title': '';
+      return FlutterI18n.translate(context, translateKey);
+    }
+    return key;
+  }
 
   bool get readOnly => forceDisabled ? true : jsonSchema.readOnly;
 
@@ -44,15 +59,19 @@ class FlutterJsonSchemaFormField extends StatelessWidget {
 
   final Function(dynamic)? onChanged;
 
+  final bool translateField;
+
+  final String translateBase;
+
+  final String fieldName;
+
   _onChanged(changeResult) {
     // field was selected at path
     // We should tell the higher level controller that we have selected a field
-
-    final correspondingTextEditingController =
-        accessValue(path, editingControllerMapping);
-
-    if (correspondingTextEditingController is TextEditingController) {
-      correspondingTextEditingController.text = changeResult ?? "";
+    final correspondingTextEditingController = accessValue(path, editingControllerMapping);
+    if ((correspondingTextEditingController is TextEditingController) && (changeResult != null)) {
+      correspondingTextEditingController.text = changeResult.toString();
+      jsonSchema.default_ = null;
     }
 
     if (onChanged != null) {
@@ -60,11 +79,30 @@ class FlutterJsonSchemaFormField extends StatelessWidget {
     }
   }
 
+  _setValue(newValue) {
+    final correspondingTextEditingController = accessValue(path, editingControllerMapping);
+    if ((correspondingTextEditingController is TextEditingController) && (newValue != null)) {
+      correspondingTextEditingController.text = newValue.toString();
+    }
+  }
+
+  get _enumToSelectNumberItems => jsonSchema.enum_
+      ?.map(
+        (e) => DropdownMenuItem<num>(
+          child: Text(e.toString()),
+          value: e,
+          onTap: () {},
+        ),
+      ).toList();
+
   get _enumToSelectItems => jsonSchema.enum_
       ?.map(
         (e) => DropdownMenuItem<String>(
-          child: Text(e),
-          value: e as String,
+          child:
+            translateField && (num.tryParse(e.toString()) == null)
+            ? I18nText(jsonSchema.title != null ? jsonSchema.title.toString() + '_enum.' + e.toString() : e.toString())
+            : Text(e.toString()),
+          value: e.toString(),
           onTap: () {},
         ),
       )
@@ -92,25 +130,25 @@ class FlutterJsonSchemaFormField extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    switch (jsonSchema.type) {
+     var translatedTitle = getTitleTranslation(context, key);
+     if (jsonSchema.default_ != null) {
+       _setValue(jsonSchema.default_);
+     }
+     var fieldValue = accessValue(path, formState);
+     var textController = accessValue(path, editingControllerMapping);
+     switch (jsonSchema.type) {
       case JsonSchemaType.string:
         if (jsonSchema.enum_ != null) {
           return Row(
             children: [
-              Text("${jsonSchema.title ?? ""} :"),
+              Text(translatedTitle),
               const SizedBox(width: 12),
               Padding(
                 padding: const EdgeInsets.all(4.0),
                 child: DropdownButton<String>(
-                  value: accessValue(
-                    path,
-                    formState,
-                  ) is String
-                      ? accessValue(
-                          path,
-                          formState,
-                        )
-                      : null,
+                  value: fieldValue is String
+                      ? fieldValue
+                      : jsonSchema.default_,
                   onChanged: _onChanged,
                   items: _enumToSelectItems,
                 ),
@@ -118,56 +156,92 @@ class FlutterJsonSchemaFormField extends StatelessWidget {
             ],
           );
         }
-
         return Padding(
           padding: const EdgeInsets.symmetric(vertical: 4),
           child: TextField(
             decoration: InputDecoration(
-              labelText: title,
+              labelText: translatedTitle
             ),
             onChanged: (value) {
               // controller.updateValue(path, value);
             },
-            controller: accessValue(path, editingControllerMapping),
+            controller: textController,
             readOnly: readOnly,
             enabled: enabled,
           ),
         );
       case JsonSchemaType.number:
         if (jsonSchema.enum_ != null) {
-          return Padding(
-            padding: const EdgeInsets.all(4.0),
-            child: DropdownButton<num>(
-              value: accessValue(
-                path,
-                formState,
-              ) is String
-                  ? accessValue(
-                      path,
-                      formState,
-                    )
-                  : null,
-              onChanged: _onChanged,
-              items: _enumToSelectItems,
-            ),
-          );
+          return Row(
+            children: [
+              Text(translatedTitle),
+              const SizedBox(width: 12),
+              Padding(
+                padding: const EdgeInsets.all(4.0),
+                child:
+                  DropdownButton<num>(
+                    value: fieldValue is String
+                        ? num.tryParse(fieldValue) ?? num.parse(fieldValue)
+                        : jsonSchema.default_,
+                    onChanged: _onChanged,
+                    items: _enumToSelectNumberItems,
+                  ),
+          )
+          ]);
         }
+
         return Padding(
           padding: const EdgeInsets.symmetric(vertical: 4),
           child: TextField(
             decoration: InputDecoration(
-              labelText: title,
+              labelText: translatedTitle,
             ),
             keyboardType: TextInputType.number,
+            inputFormatters: <TextInputFormatter>[
+              FilteringTextInputFormatter.digitsOnly
+            ],
             onChanged: (value) {
               // controller.updateValue(path, value);
             },
-            controller: accessValue(path, editingControllerMapping),
+            controller: textController,
             readOnly: readOnly,
             enabled: enabled,
           ),
         );
-      case JsonSchemaType.object:
+       case JsonSchemaType.boolean:
+         return Row(
+             children: [
+               Text(translatedTitle),
+               const SizedBox(width: 12),
+               Padding(
+                 padding: const EdgeInsets.all(4.0),
+                 child: DropdownButton<bool>(
+                   value: fieldValue is String
+                       ? (fieldValue.toLowerCase() == 'true')
+                       : jsonSchema.default_,
+                   onChanged: _onChanged,
+                   items: [
+                     DropdownMenuItem<bool>(
+                       child:
+                          translateField
+                          ? I18nText('template.options.true')
+                          : const Text('true'),
+                       value: true,
+                       onTap: () {},
+                     ),
+                     DropdownMenuItem<bool>(
+                       child:
+                       translateField
+                           ? I18nText('template.options.false')
+                           : const Text('false'),
+                       value: false,
+                       onTap: () {},
+                     ),
+                   ],
+                 ),
+               )
+             ]);
+       case JsonSchemaType.object:
         final svgProps = ['name', 'size', 'type', 'lastModified', 'data'];
         if (listEquals(svgProps, jsonSchema.properties.keys.toList())) {
           return fileWidget ?? Container();
@@ -219,6 +293,8 @@ class FlutterJsonSchemaFormField extends StatelessWidget {
                   disabled: forceDisabled,
                   onChanged: onChanged,
                   formState: formState,
+                  translate: translateField,
+                  translateBase: translateBase != '' ? translateBase + '.' + fieldName : jsonSchema.title.toString()
                 ),
               )
             ],